- redesign of customer standard model

+++ design/views.py
@@ -1,5 +1,6 @@
 import json
-
+import re
+import uuid as uuid_id
 from django.contrib.auth import authenticate, login as django_login, logout as django_logout, update_session_auth_hash
 from django.contrib.auth.forms import PasswordChangeForm
 from django.db.models import Q
@@ -25,8 +26,8 @@ from rest_framework.views import APIView
 
 product_json_dic = {
     "flyer":"합판 전단지","card":"명함","envelope":"봉투","banner":"배너","ticket":"상품권","roll":"현수막/족자",
-    "real":"실사 출력","octopus":"합판 문어발","moon":"문고리","sticker":"스티커", "ncr":"양식지/NCR지",
-    "magnet":"종이자석","catalog":"카달로그/브로셔","postit":"포스트잇/떡메모지","advertising":"판촉/홍보","bee":"비품"
+    "real":"실사 출력","octopus":"합판 문어발","moon":"문고리","sticker":"스티커", "ncr":"NCR지","ncr2":"양식지",
+    "magnet":"종이자석","etc":"기타 인쇄","postit":"포스트잇/떡메모지","advertising":"판촉/홍보","bee":"비품"
  }
 
 class Login(generic.FormView):
@@ -146,23 +147,43 @@ class ProductView(ProfileMixin, generic.FormView, generic.ListView):
         return self
 
 
-class ProductSampleView(ProfileMixin, generic.FormView, generic.ListView):
+class ProductSampleView(viewmixin.PageableMixin, ProfileMixin, generic.FormView, generic.ListView):
     template_name = 'design/product_sample.html'
     context_object_name = 'sample_list'
+    paginate_by = 5
     form_class = forms.ProductForm
 
     def get_queryset(self):
         product_category = self.request.GET.get('item')
         sector_category = self.request.GET.get('sector')
-        return managing_model.Sample.objects.select_related( 'sectors_category__parent').filter(state=True)
+        return managing_model.Sample.objects.select_related( 'sectors_category__parent').order_by( '-created').filter(state=True)
 
     def get_context_data(self, **kwargs):
         context = super(ProductSampleView, self).get_context_data(**kwargs)
         product_category = self.request.GET.get('item')
         sector_category = self.request.GET.get('sector')
         context['item'] = product_category
-        # context['menu_slug']= self.kwargs['menu_slug']
-        # context['sector_slug'] = self.kwargs['sector_slug']
+        if self.request.user.is_authenticated:
+            cart = design_model.CartProduct.objects.filter(user=self.request.user)
+            cart_button_list = []
+            num = 0
+            for i in cart:
+                text = json.loads(i.json_text)
+                name = text['title']+" "+ text['size']
+                if '단면' in text['side']:
+                    front = name + " 앞"
+                    data = [num, front]
+                    cart_button_list.append(data)
+                elif '양면' in text['side']:
+                    front = name + " 앞"
+                    back = name + " 뒤"
+                    data = [num, front]
+                    cart_button_list.append(data)
+                    num += 1
+                    data = [num, back]
+                    cart_button_list.append(data)
+                num += 1
+            context['cart_button_list'] = cart_button_list
         return context
 
     def form_valid(self, form):
@@ -172,7 +193,8 @@ class ProductSampleView(ProfileMixin, generic.FormView, generic.ListView):
 class ProductConfirmView(ProfileMixin, generic.FormView, generic.ListView):
     template_name = 'design/product_confirm.html'
     context_object_name = 'confirm_list'
-    form_class = forms.ProductForm
+    form_class = forms.OrderProfileForm
+    success_url = ''
 
     def get_queryset(self):
         product_category = self.request.GET.get('item')
@@ -181,12 +203,71 @@ class ProductConfirmView(ProfileMixin, generic.FormView, generic.ListView):
 
     def get_context_data(self, **kwargs):
         context = super(ProductConfirmView, self).get_context_data(**kwargs)
+        if self.request.user.is_authenticated:
+            customer_profile = member_model.CustomerProfile.objects.get(user=self.request.user)
+            form = self.form_class(instance=customer_profile)  # get_or_create 는 신규 생성인지 boolean 튜플 제공
+            context['form'] = form
         product_category = self.request.GET.get('item')
         sector_category = self.request.GET.get('sector')
         context['item'] = product_category
         # context['menu_slug']= self.kwargs['menu_slug']
         # context['sector_slug'] = self.kwargs['sector_slug']
         return context
+
+    def form_valid(self, form):
+        self.success_url = self.request.META.get('HTTP_REFERER')
+        form.save(commit=False)
+        # 금액 검수
+        price_model = design_model.ProductPriceAPI.objects.all()
+        i = design_model.CartProduct.objects.values('json_text')
+        success_list = []
+        error_list = []
+        for i in i:
+            d = json.loads(i['json_text'])
+            d['deal'] = d['deal'].replace(',', '')
+            d['deal'] = re.findall("\d+", d['deal'])[0]
+            model_model = price_model.filter(size=d['size'],paper=d['paper'],side=d['side'],deal=d['deal'])
+            price = ''
+            for i in model_model:
+                price = i.sell
+            price2 = float(d['price'])
+            if price2 == price:
+                success_list.append(d)
+            else:
+                error_list.append(d)
+                design_model.CartPriceProblem.objects.create(user=self.request.user,json_text=d)
+        if error_list:
+            messages.error(self.request, '장바구니에 비정상적인 금액을 감지하였습니다. 관리자에게 문의주세요')
+            return redirect(self.success_url)
+        if not success_list:
+            messages.error(self.request, '장바구니가 비어있습니다.')
+            return redirect(self.success_url)
+
+        # order_info
+        uuid = uuid_id.uuid4().hex[:10]
+        user = self.request.user
+        dic = form.cleaned_data
+        dic['user'] = user
+        dic['uuid'] = uuid
+        if dic['address']:
+            im_user = member_model.CustomerProfile.objects.filter(user=self.request.user, address__icontains= dic['address'])
+            if not im_user:
+                dic['address_confirm'] = False
+        else:
+            return redirect(self.success_url)
+
+        data_model = design_model.CustomerOrderInfo.objects.create(**dic)
+        # order_product
+        for data in success_list:
+            data['sell'] = data.pop('price')
+            print(data)
+            data['customer_order_info'] = data_model
+            design_model.CustomerOrderProduct.objects.create(**data)
+        # clean_cart_product
+        cart = design_model.CartProduct.objects.filter(user=self.request.user)
+        cart.delete()
+
+        return redirect(self.success_url)
 
 
 class ProfileView(ProfileMixin,generic.TemplateView):
@@ -198,13 +279,54 @@ class FaqView(ProfileMixin,generic.TemplateView):
 class NewsView(ProfileMixin,generic.TemplateView):
     template_name = 'design/news.html'
 
+class OrderListView(ProfileMixin, generic.ListView):
+    template_name = 'design/order_list.html'
+    model = design_model.CustomerOrderInfo
+    context_object_name = 'order_list'
+    paginate_by = 10
+
+    def get_queryset(self):
+        if self.request.user.is_authenticated:
+            queryset = super().get_queryset().prefetch_related('customer_order_product').filter(user=self.request.user).order_by('-created')
+        else:
+            queryset = ''
+        return queryset
+
+    def get_context_data(self, **kwargs):
+        context = super(OrderListView, self).get_context_data(**kwargs)
+        if self.request.user.is_authenticated:
+            for z in context['order_list']:
+                data = z.customer_order_product.all()
+                name = ''
+                total = 0
+                count = 0
+                for x in data:
+                    if count == 0:
+                        name = x.size
+                        z.title = x.title
+                    total += x.sell
+                    count += 1
+                z.total = total
+                if count == 0:
+                    z.product = '내역이 없습니다.'
+                elif count == 1:
+                    z.product = name
+                else:
+                    z.product = name +' 외 '+ str(count-1)+' 건'
+        return context
+
 class JoinView(ProfileMixin, generic.FormView):
     template_name = 'design/join.html'
     form_class = forms.CreateUserForm
     success_url = "/design/join/"
 
     def form_valid(self, form):
-        print(form.cleaned_data)
+        username = form.cleaned_data['username']
+        password = form.cleaned_data['password1']
+        user = User.objects.create_user(username=username, password=password)
+        if user:
+            django_login(self.request, user, backend='django.contrib.auth.backends.ModelBackend')
+            return redirect('/design/my_page/')
         response = super(JoinView, self).form_valid(form)
         return response
 
@@ -218,6 +340,7 @@ class MyPageView(ProfileMixin, generic.FormView):
         user = self.request.user.id
         if user:
             customer_profile = member_model.CustomerProfile.objects.get_or_create(user=self.request.user)
+            managing_model.CustomerMemo.objects.get_or_create(customer=self.request.user)
             form = self.form_class(instance=customer_profile[0]) # get_or_create 는 신규 생성인지 boolean 튜플 제공
             context['form'] = form
             password_form = PasswordChangeForm(self.request.user)
@@ -240,23 +363,29 @@ class AjaxPriceView(APIView):
     def post(self, request, format=None):
         serializer = serializers.PaymentSerializer(data=request.data)
         if serializer.is_valid():
-            size= request.data['size']
-            paper = request.data['paper']
-            side = request.data['side']
-            deal = request.data['deal']
+            size= str(request.data['size'])
+            paper = str(request.data['paper'])
+            side = str(request.data['side'])
+            deal = str(request.data['deal'])
+            deal = deal.replace(',', '')
+            deal = re.findall("\d+", deal)[0]
             img_models = design_model.ProductImg.objects.all()
-            size_img = img_models.filter(name__icontains = size)
+            size_img = img_models.filter(name = size)
+            paper_img = img_models.filter(name = paper)
+            data = design_model.ProductPriceAPI.objects.filter(size=size,paper=paper,side=side,deal=deal)
+            sell = [i.sell for i in data]
+            if sell:
+                price = sell
+            else:
+                price = 0
             if size_img:
                 size_img = [i.images.url for i in size_img]
-            else:
-                size_img = "default"
-
+            if paper_img:
+                paper_img = [i.images.url for i in paper_img]
             back_dic = {
-                'sell_price':'50,000',
-                'sell_tax': '5,000',
-                'sell_total': '55,000',
+                'sell_price':price,
                 'size_img':size_img,
-                'paper_img':'http://www.aceprinting.co.kr/skin/guide/images/goods/Goods_04_01_03.jpg'
+                'paper_img':paper_img
             }
             return Response(back_dic)
         else:
@@ -285,13 +414,16 @@ class CartProductView(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, format=None):
-        data_id = request.data['num']
-        query = design_model.CartProduct.objects.get(id=data_id, user=request.user)
-        if query:
-            query.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.data['num']:
+            data_id = request.data['num']
+            query = design_model.CartProduct.objects.get(id=data_id, user=request.user)
+            if query:
+                query.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            else:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
 class SampleListView(viewmixin.PageableMixin, viewmixin.DataSearchFormMixin,
@@ -303,7 +435,7 @@ class SampleListView(viewmixin.PageableMixin, viewmixin.DataSearchFormMixin,
     context_object_name = 'sample_list'
 
     def get_queryset(self):
-        queryset = super().get_queryset().order_by( '-created')
+        queryset = super().get_queryset().order_by( '-created').filter(state=True)
         form = self.data_search_form(self.request.GET)
         if form.is_valid() and form.cleaned_data['q']:
             q = form.cleaned_data['q']
